backend/requirements/dev.txt
-r base.txt
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.1
mypy==1.11.1
django-stubs==5.0.4
djangorestframework-stubs==3.15.0
mypy-extensions==1.0.0
django-extensions==3.2.3
ipython==8.26.0
